[refactor](be) Localize parquet map repeated assembly
### What problem does this PR solve?

Issue Number: close #xxx

Related PR: #xxx

Problem Summary: Localize MAP-only repeated level assembly and value stream helpers in the new parquet MapColumnReader after LIST Dremel assembly was inlined, reducing the public nested reader helper surface. Update the complex column refactor document with the current Phase 3 and Phase 4 status.

### Release note

None

### Check List (For Author)

- Test: Manual test
    - git diff --check
- Behavior changed: No
- Does this need documentation: No
